requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.9
django-filter==24.3
django-cors-headers==4.4.0
python-decouple==3.8
celery==5.4.0
redis==5.0.8
requests==2.32.3
django-rest-knox==4.2.0  # For token-based authentication
django-phonenumber-field==7.3.0
phonenumbers==8.13.47
geopy==2.4.1  # For geolocation
